Changed routes so data from welcome controller posts to first page/start controller. Added a link to first page on second page so you can go back to step_one/:token and edit data entered in welcome controller. The begin/continue application button on step_one/:token however doesnt work now, need to fix the routing for that to advance to step_two.
