+++ app/controllers/welcome_controller.rb
@@ -4,26 +4,16 @@ class WelcomeController < ApplicationController
     @apprenticeform = ApprenticeForm.new
   end
 
-  def create
-    @apprenticeform = ApprenticeForm.create(apprentice_form_params)
+  # def update
+  #   @apprenticeform = ApprenticeForm.find_by_token(params[:id])
+  #   @apprenticeform.update_attributes(apprentice_form_params)
+  #   redirect_to edit_second_step_path
+  # end
 
-    if @apprenticeform.errors.any?
-      render :edit
-    else
-      redirect_to edit_second_step_path(@apprenticeform.token)
-    end
-  end
-
-  def update
-    @apprenticeform = ApprenticeForm.find_by_token(params[:id])
-    @apprenticeform.update_attributes(apprentice_form_params)
-    redirect_to edit_second_step_path
-  end
-
-  private
-  def apprentice_form_params
-    params.require(:apprentice_form).permit(:first_name, :last_name, :email, :start_date, :end_date, :willing_to_move, :id)
-  end
+  # private
+  # def apprentice_form_params
+  #   params.require(:apprentice_form).permit(:first_name, :last_name, :email, :start_date, :end_date, :willing_to_move, :id)
+  # end
 
 end
