+++ app/helpers/start_helper.rb
@@ -1,0 +1,2 @@
+module StartHelper
+end
