+++ app/views/demographics/show.html.erb
@@ -14,4 +14,6 @@
   <%= f.hidden_field :id %>
   <%= f.submit "Next Step", :class => 'button' %>
 <% end %>
+
+<%= link_to "Link to first page", edit_first_step_path%>
 </div>
