+++ app/views/start/_footer.html.erb
@@ -1,0 +1,4 @@
+<footer>
+  <a href="http://www.seesparkbox.com" title="Sparkbox logo"><%= image_tag "SBX_logo_icon_2.svg", class: "logo-footer", alt: "Sparkbox Logo" %></a>
+  <p>Proudly <a href="https://github.com/sparkbox/apprentices-page">built</a> by <a href="http://www.seesparkbox.com">Sparkbox</a> Apprentices v3.0</p>
+</footer>
