+++ app/views/start/show.html.erb
@@ -3,11 +3,8 @@
 </header>
 <%= image_tag "SBX_logo_icon_2.svg", class: "logo-special", alt: "Sparkbox Logo" %>
 
-<%= render 'apprenticeship_information' %>
-
 <section class="container">
-  <h1 class="info-header">Time to Apply</h1>
-  <p class="info-secondary">How bad do you want it? What you provide us is very important — this is your opportunity to present your case. Choose your answers wisely, my friend!</p>
+
   <%= form_for @apprenticeform, :url => step_one_path, method: :post, html: {class:"apprentice_application_form"} do |f| %>
   <%= render 'basic_information_step', :f => f %>
 
@@ -16,7 +13,7 @@
   <% end %>
 
 
-  <!-- <%= link_to "step 2", edit_second_step_path('asdftoken')%> -->
+  <%= link_to "step 2", edit_second_step_path('asdftoken')%>
 
   <%= debug @apprenticeform if Rails.env.development? %>
