+++ test/controllers/start_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class StartControllerTest < ActionController::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
