+++ test/helpers/start_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class StartHelperTest < ActionView::TestCase
+end
